feat(onchain_endpoint): added components fetching from DB
