+++ infra/pragma-node/config/.env.example
@@ -7,5 +7,6 @@ PORT=3000
 KAFKA_BROKERS="pragma-kafka:9092"
 
 # Provider
+# or https://free-rpc.nethermind.io/mainnet-juno
 RPC_URL=https://free-rpc.nethermind.io/sepolia-juno
 NETWORK=testnet
