+++ pragma-node/src/handlers/entries/mod.rs
@@ -71,11 +71,11 @@ pub struct GetVolatilityResponse {
 
 #[derive(Debug, Serialize, Deserialize, ToSchema)]
 pub struct OnchainEntry {
-    publisher: String,
-    source: String,
-    price: String,
-    tx_hash: String,
-    timestamp: u64,
+    pub publisher: String,
+    pub source: String,
+    pub price: String,
+    pub tx_hash: String,
+    pub timestamp: u64,
 }
 
 #[derive(Debug, Serialize, Deserialize, ToSchema)]
@@ -86,8 +86,7 @@ pub struct GetOnchainEntryResponse {
     decimals: u32,
     nb_sources_aggregated: u32,
     asset_type: String,
-    // TODO: query the onchain repository for the components
-    // components: Vec<OnchainEntry>,
+    components: Vec<OnchainEntry>,
 }
 
 /// Query parameters structs
@@ -131,21 +130,6 @@ impl Default for GetEntryParams {
             interval: Some(Interval::default()),
             routing: Some(false),
             aggregation: Some(AggregationMode::default()),
-        }
-    }
-}
-
-#[derive(Debug, Deserialize, IntoParams, ToSchema)]
-pub struct GetOnchainParams {
-    pub aggregation: Option<AggregationMode>,
-    pub timestamp: Option<u64>,
-}
-
-impl Default for GetOnchainParams {
-    fn default() -> Self {
-        Self {
-            aggregation: Some(AggregationMode::default()),
-            timestamp: Some(chrono::Utc::now().timestamp_millis() as u64),
         }
     }
 }
